==> 5. Model Stacking.ipynb <==
# Model Stacking

In this notebook, what I will do is to take a linear combination between the result of the the models and try to see if we can get a high score out of this method.

import numpy as np
import pandas as pd

cat = pd.read_csv('submission/result_cat.csv')
rf = pd.read_csv('submission/result_rf.csv')
print("Finished Loading the prediction results.")

for weight in np.linspace(0.1, 0.9, 9):
    print('Now start stacking for weight ' + str(np.round(weight,1)))
    stack = pd.DataFrame()
    stack['ParcelId'] = lgb['ParcelId']
    for col in ['201610', '201611', '201612', '201710', '201711', '201712']:
        stack[col] = weight * rf[col] + (1 - weight) * lgb[col]
    stack.to_csv(f'submission/result_stack_{weight}.csv', index=False)
print("Finished Stacking.")

It turns out the best result is the stacking result with the following linear combination.


$$ Best = 0.6 \times CatBoost + 0.4 \times RandomForest $$
